`openProject` > `renderProject` and seperate
