+++ app/project_template.html
@@ -27,21 +27,11 @@
 					<th>Rechnungsnummer:</th>
 					<th>Betrag in €:</th>
 				</tr>
-				<tr>
-					<td>Reparatur</td>
-					<td>001</td>
-					<td>950</td>
-				</tr>
-				<tr>
-					<td>Treibstoff</td>
-					<td>002</td>
-					<td>110</td>
-				</tr>
 			</table>
 		</div>
 		<div class="col-footer">
 			<button class="btn btn-primary">Hinzufügen</button>
-			<span id="material-costs-display" class="costs-display">1060€</span>
+			<span id="material-costs-display" class="costs-display"></span>
 		</div>
 	</div>
 
@@ -54,16 +44,11 @@
 					<th>Stunden:</th>
 					<th></th>
 				</tr>
-				<tr>
-					<td>Baggerfahrer <span class="align-right">9,50€</span></td>
-					<td>16</td>
-					<td><button class="btn btn-primary btn-small">&plus;</button></td>
-				</tr>
 			</table>
 		</div>
 		<div class="col-footer">
 			<button class="btn btn-primary">Typ hinzufügen</button>
-			<span id="wages-costs-display" class="costs-display">152€</span>
+			<span id="wages-costs-display" class="costs-display"></span>
 		</div>
 	</div>
 
@@ -74,15 +59,15 @@
 		<div class="section-bottom">
 			<div class="section-bottom">
 				<label>Brutto:</label>
-				<input type="text" id="brutto-bill-input" value="2409,50">€
+				<input type="text" id="brutto-bill-input">€
 				<br><br>
 				<label>Netto:</label>
-				<span>1999,89€</span>
+				<span>0,00€</span>
 				<hr>
 				<label>Bilanz:</label>
 				<div id="bilanz-display">
-						<span>520,89€</span><br>
-						<span>26,0%</span>
+						<span>0,00€</span><br>
+						<span>0,0%</span>
 				</div>
 		</div>
 		</div>
